Move engine-specific code into Engine, and move into its own file
