+++ src/main.ts
@@ -1,38 +1,16 @@
-import { Args, Engine, getTasks, Task } from "grimoire-kolmafia";
-import { args, daily, fmt, halloween, printArgs } from "./util";
+import { Args, getTasks } from "grimoire-kolmafia";
+import { print, wait } from "kolmafia";
+
+import { args, halloween, printArgs } from "./util";
 import { farm } from "./farm";
 import { diet } from "./diet";
-import { print, totalTurnsPlayed, wait } from "kolmafia";
 import { pvp } from "./pvp";
-import { get } from "libram";
 import { autoscend, pathQuest } from "./paths";
-
-class HalfloopEngine extends Engine {
-  turns: Map<string, number[]> = new Map();
-
-  execute(task: Task<never>): void {
-    const startingTurns = totalTurnsPlayed();
-    super.execute(task);
-    const oldTurns = this.turns.get(task.name);
-    this.turns.set(task.name, [
-      totalTurnsPlayed() - startingTurns,
-      ...(oldTurns ?? []),
-    ]);
-  }
-
-  printTurns(): void {
-    print("Tasks Turns:");
-    for (const [name, turns] of this.turns.entries()) {
-      print(`- ${name}: ${turns.reduce((a, b) => a + b)} (${turns})`);
-    }
-  }
-}
+import { HalfloopEngine } from "./engine";
 
 export function main(command = ""): void {
   Args.fill(args, command);
 
-  const startingTurns = totalTurnsPlayed();
-  const startingSwagger = get("availableSwagger");
 
   const tasks = getTasks([pvp, autoscend, pathQuest(), diet, farm]);
   const engine = new HalfloopEngine(tasks);
@@ -73,42 +51,6 @@ export function main(command = ""): void {
     engine.run();
   } finally {
     engine.destruct();
-    engine.printTurns();
-
-    print("");
-
-    const endingTurns = totalTurnsPlayed();
-    const endingSwagger = get("availableSwagger");
-
-    const [totalTurnsSpent, totalSwagger] = daily(({ get, set }) => {
-      set(
-        "halfloop_turnsSpent",
-        get("halfloop_turnsSpent") + (endingTurns - startingTurns),
-      );
-      set(
-        "halfloop_swagger",
-        get("halfloop_swagger") + (endingSwagger - startingSwagger),
-      );
-      return [get("halfloop_turnsSpent"), get("halfloop_swagger")];
-    });
-
-    const meat = get("garboResultsMeat", 0);
-    const item = get("garboResultsItems", 0);
-    const embezzlers = get("garboEmbezzlerCount", 0);
-    const yachtzees = get("garboYachtzeeCount", 0);
-
-    print("Final Results");
-    print(`* Total Turns Spent: ${totalTurnsSpent}`);
-    print(
-      `* Garbo Results: ${fmt(meat)} meat + ${fmt(item)} items = ${fmt(
-        meat + item,
-      )}`,
-    );
-    print(
-      `* Garbo Actions: ${fmt(embezzlers)} embezzlers and ${fmt(
-        yachtzees,
-      )} yachtzees`,
-    );
-    print(`* Swagger: ${fmt(totalSwagger)}`);
+    engine.printSummary();
   }
 }
